Moved over more examples
