+++ thruster-server/src/lib.rs
@@ -3,5 +3,3 @@ pub mod thruster_server;
 
 #[cfg(feature="hyper_server")]
 pub mod hyper_server;
-#[cfg(feature="hyper_server")]
-pub mod basic_hyper_context;
