+++ thruster/examples/hyper_most_basic.rs
@@ -1,17 +1,18 @@
 #[macro_use] extern crate thruster;
 extern crate futures_legacy;
-extern crate serde_json;
+extern crate hyper;
 
 use std::boxed::Box;
 use futures_legacy::future;
-use thruster::{App, BasicContext as Ctx, MiddlewareChain, MiddlewareReturnValue, Request};
-use thruster::builtins::server::Server;
-use thruster::builtins::query_params::query_params;
-use thruster::server::ThrusterServer;
+
+use thruster::{App, Context, MiddlewareChain, MiddlewareReturnValue};
+use thruster::hyper_server::Server;
+use thruster::thruster_context::basic_hyper_context::{generate_context, BasicHyperContext as Ctx, HyperRequest};
+use thruster::ThrusterServer;
 
 fn plaintext(mut context: Ctx, _next: impl Fn(Ctx) -> MiddlewareReturnValue<Ctx>  + Send + Sync) -> MiddlewareReturnValue<Ctx> {
-  let val = serde_json::to_string(&context.query_params).unwrap();
-  context.body(&val);
+  let val = "Hello, World!".as_bytes().to_vec();
+  context.set_body(val);
 
   Box::new(future::ok(context))
 }
@@ -19,9 +20,8 @@ fn plaintext(mut context: Ctx, _next: impl Fn(Ctx) -> MiddlewareReturnValue<Ctx>
 fn main() {
   println!("Starting server...");
 
-  let mut app = App::<Request, Ctx>::new_basic();
+  let mut app = App::<HyperRequest, Ctx>::create(generate_context);
 
-  app.use_middleware("/", middleware![Ctx => query_params]);
   app.get("/plaintext", middleware![Ctx => plaintext]);
 
   let server = Server::new(app);
